feat: скопировал dev сюда

+++ Poolit/Models/UserFile.cs
@@ -2,9 +2,9 @@
 
 public class UserFile
 {
-    public ulong Id { get; set; }
-    public ulong AuthorId { get; set; }
+    public int Id { get; set; }
+    public int AuthorId { get; set; }
     public string Name { get; set; }
     public string Url { get; set; }
-    public ulong Size { get; set; }
+    public int Size { get; set; }
 }
