+++ Poolit/Services/IFileService.cs
@@ -4,7 +4,7 @@ namespace Poolit.Services;
 
 public interface IFileService
 {
-    public string GetFileUrlById(ulong id);
+    public string GetFileUrlById(int id);
     public bool FileExists(string path);
-    public UserFile[] GetUserFiles(ulong userId);
+    public UserFile[] GetUserFiles(int userId);
 }
